requirements.txt
crewai
langchain-groq
streamlit
python-dotenv
langchain-google-genai
google-generativeai
duckduckgo-search
langchain-community
